Always launch the music thread
This feels a little controversial (because I suspect the music functions are used infrequently across all Immediate2D use cases), but there is a cold-startup time associated with spinning up the thread and loading the Windows MIDI synth.

So our choices are to advise users to send a short dummy note at startup (for those users that are actually interested in the PlayMusic function), don't say anything and have users experience a clipped first note, or always launch a second thread that basically sleeps forever.

The latter seemed like the least bad of the three.